image-ai/requirements.txt
google-genai
opencv-python
pillow
numpy